pyproject.toml
[tool.poetry]
name = "auto-EPU"
version = "0.1.0"
description = "Utilzing LLM to automatically construct economic textual indicators."
authors = ["Hsiu-Hsuan Yeh"]
license = "MIT"
readme = "README.md"
packages = [{include = "auto_EPU"}]


[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
orjson = "^3.9.12"
tqdm = "^4.66.1"
pandas = "^2.2.0"
mlflow = "^2.10.2"
openpyxl = "^3.1.2"
langchain = "^0.1.7"
ckip-transformers = "^0.3.4"
scikit-learn = "^1.4.1.post1"
sentence-transformers = "^2.3.1"
keybert = "^0.8.4"
hvplot = "^0.9.2"
bokeh = "^3.3.4"
wordcloud = "^1.9.3"
matplotlib = "^3.8.3"
selenium = "^4.17.2"
geckodriver = "^0.0.1"
tiktoken = "<0.6.0"
seaborn = "^0.13.2"
plotly = "^5.20.0"
pyzmq = "26.0.3"
llm-research = "0.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
jupyter-bokeh = "^3.0.7"
hvplot = "^0.9.1"
pytest = "^7.4.4"
jupyterlab-execute-time = "2.3.1"
jupyterlab = "3.6.6"
jupyterlab-lsp = "4.3.0"
python-lsp-server = {extras = ["all"], version = "^1.8.2"}
poetry-kernel = "^0.1.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
